Split country selector into separate files with shared hook
