--- app/components/layout/country-selector.tsx
@@ -1,305 +1,0 @@
-import { CaretDownIcon, CheckCircleIcon } from "@phosphor-icons/react";
-import * as Popover from "@radix-ui/react-popover";
-import { CartForm } from "@shopify/hydrogen";
-import type { CartBuyerIdentityInput } from "@shopify/hydrogen/storefront-api-types";
-import { useThemeSettings } from "@weaverse/hydrogen";
-import { useEffect, useRef } from "react";
-import ReactCountryFlag from "react-country-flag";
-import { useInView } from "react-intersection-observer";
-import {
-  useFetcher,
-  useLocation,
-  useRouteLoaderData,
-  useSubmit,
-} from "react-router";
-import { ScrollArea } from "~/components/scroll-area";
-import type { RootLoader } from "~/root";
-import type { I18nLocale, Localizations } from "~/types/others";
-import type { ThemeSettings } from "~/types/weaverse";
-import { DEFAULT_LOCALE } from "~/utils/const";
-
-export function CountrySelector() {
-  const fetcher = useFetcher();
-  const submit = useSubmit();
-  const rootData = useRouteLoaderData<RootLoader>("root");
-  const selectedLocale = rootData?.selectedLocale ?? DEFAULT_LOCALE;
-  const { pathname, search } = useLocation();
-  const pathWithoutLocale = `${pathname.replace(
-    selectedLocale.pathPrefix,
-    "",
-  )}${search}`;
-
-  const countries = (fetcher.data ?? {}) as Localizations;
-  const defaultLocale = countries?.default;
-  const defaultLocalePrefix = defaultLocale
-    ? `${defaultLocale?.language}-${defaultLocale?.country}`
-    : "";
-
-  const { ref, inView } = useInView({
-    threshold: 0,
-    triggerOnce: true,
-  });
-
-  const observerRef = useRef(null);
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation> --- IGNORE ---
-  useEffect(() => {
-    ref(observerRef.current);
-  }, [ref, observerRef]);
-
-  // Get available countries list when in view
-  useEffect(() => {
-    if (!inView || fetcher.data || fetcher.state === "loading") {
-      return;
-    }
-    fetcher.load("/api/countries");
-  }, [inView, fetcher]);
-
-  function handleLocaleChange({
-    redirectTo,
-    buyerIdentity,
-  }: {
-    redirectTo: string;
-    buyerIdentity: CartBuyerIdentityInput;
-  }) {
-    submit(
-      {
-        redirectTo,
-        cartFormInput: JSON.stringify({
-          action: CartForm.ACTIONS.BuyerIdentityUpdate,
-          inputs: { buyerIdentity },
-        }),
-      },
-      { method: "POST", action: "/cart" },
-    );
-  }
-
-  return (
-    <div ref={observerRef} className="grid w-80 gap-4">
-      <Popover.Root>
-        <Popover.Trigger asChild>
-          <button
-            type="button"
-            className="flex w-full cursor-pointer items-center gap-2 overflow-clip rounded-md border border-line-subtle px-4 py-3 text-left outline-hidden"
-            aria-label="Select country"
-          >
-            <ReactCountryFlag
-              svg
-              countryCode={selectedLocale.country}
-              className="rounded-xs"
-              style={{ width: "24px", height: "14px" }}
-            />
-            <span>{selectedLocale.label}</span>
-            <CaretDownIcon className="ml-auto h-4 w-4" />
-          </button>
-        </Popover.Trigger>
-        <Popover.Portal>
-          <Popover.Content>
-            <ScrollArea
-              size="sm"
-              style={{ maxHeight: 160, width: 320 }}
-              rootClassName="my-2 rounded-lg bg-neutral-800 py-2"
-            >
-              {countries &&
-                Object.keys(countries).map((countryPath) => {
-                  const countryLocale = countries[countryPath];
-                  const isSelected =
-                    countryLocale.language === selectedLocale.language &&
-                    countryLocale.country === selectedLocale.country;
-
-                  return (
-                    <Popover.Close
-                      aria-label={`Select ${countryLocale.label} country`}
-                      key={countryPath}
-                      type="button"
-                      onClick={() =>
-                        handleLocaleChange({
-                          redirectTo: getCountryUrlPath({
-                            countryLocale,
-                            defaultLocalePrefix,
-                            pathWithoutLocale,
-                          }),
-                          buyerIdentity: {
-                            countryCode: countryLocale.country,
-                          },
-                        })
-                      }
-                      className="flex w-full cursor-pointer items-center gap-2 bg-neutral-800 p-2 px-4 py-2 text-left text-white transition hover:bg-neutral-600"
-                    >
-                      <ReactCountryFlag
-                        svg
-                        countryCode={countryLocale.country}
-                        className="rounded-xs"
-                        style={{ width: "24px", height: "14px" }}
-                      />
-                      <span>{countryLocale.label}</span>
-                      {isSelected ? (
-                        <span className="ml-auto">
-                          <CheckCircleIcon className="h-5 w-5" />
-                        </span>
-                      ) : null}
-                    </Popover.Close>
-                  );
-                })}
-            </ScrollArea>
-          </Popover.Content>
-        </Popover.Portal>
-      </Popover.Root>
-    </div>
-  );
-}
-
-function getCountryUrlPath({
-  countryLocale,
-  defaultLocalePrefix,
-  pathWithoutLocale,
-}: {
-  countryLocale: I18nLocale;
-  pathWithoutLocale: string;
-  defaultLocalePrefix: string;
-}) {
-  let countryPrefixPath = "";
-  const countryLocalePrefix = `${countryLocale.language}-${countryLocale.country}`;
-  if (countryLocalePrefix !== defaultLocalePrefix) {
-    countryPrefixPath = `/${countryLocalePrefix.toLowerCase()}`;
-  }
-  return `${countryPrefixPath}${pathWithoutLocale}`;
-}
-
-export function HeaderCountrySelector() {
-  const fetcher = useFetcher();
-  const submit = useSubmit();
-  const rootData = useRouteLoaderData<RootLoader>("root");
-  const selectedLocale = rootData?.selectedLocale ?? DEFAULT_LOCALE;
-  const { countryNameDisplay } = useThemeSettings<ThemeSettings>();
-  const { pathname, search } = useLocation();
-  const pathWithoutLocale = `${pathname.replace(
-    selectedLocale.pathPrefix,
-    "",
-  )}${search}`;
-
-  const countries = (fetcher.data ?? {}) as Localizations;
-  const defaultLocale = countries?.default;
-  const defaultLocalePrefix = defaultLocale
-    ? `${defaultLocale?.language}-${defaultLocale?.country}`
-    : "";
-
-  const { ref, inView } = useInView({
-    threshold: 0,
-    triggerOnce: true,
-  });
-
-  const observerRef = useRef(null);
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation> --- IGNORE ---
-  useEffect(() => {
-    ref(observerRef.current);
-  }, [ref, observerRef]);
-
-  // Get available countries list when in view
-  useEffect(() => {
-    if (!inView || fetcher.data || fetcher.state === "loading") {
-      return;
-    }
-    fetcher.load("/api/countries");
-  }, [inView, fetcher]);
-
-  function handleLocaleChange({
-    redirectTo,
-    buyerIdentity,
-  }: {
-    redirectTo: string;
-    buyerIdentity: CartBuyerIdentityInput;
-  }) {
-    submit(
-      {
-        redirectTo,
-        cartFormInput: JSON.stringify({
-          action: CartForm.ACTIONS.BuyerIdentityUpdate,
-          inputs: { buyerIdentity },
-        }),
-      },
-      { method: "POST", action: "/cart" },
-    );
-  }
-
-  return (
-    <div className="hidden md:flex items-center gap-1.5" ref={observerRef}>
-      <Popover.Root>
-        <Popover.Trigger asChild>
-          <button
-            type="button"
-            className="flex cursor-pointer items-center gap-1 p-1.5 outline-hidden"
-            aria-label="Select country"
-          >
-            <ReactCountryFlag
-              svg
-              countryCode={selectedLocale.country}
-              className="rounded-xs"
-              style={{ width: "24px", height: "16px" }}
-            />
-            {countryNameDisplay === "full" ? (
-              <>
-                <span className="hidden xl:inline">{selectedLocale.label}</span>
-                <span className="xl:hidden">{selectedLocale.country}</span>
-              </>
-            ) : (
-              <span>{selectedLocale.country}</span>
-            )}
-            <CaretDownIcon className="h-3 w-3" />
-          </button>
-        </Popover.Trigger>
-        <Popover.Portal>
-          <Popover.Content align="center" sideOffset={8} className="z-50">
-            <ScrollArea
-              size="sm"
-              style={{ maxHeight: 240, width: 256 }}
-              rootClassName="border border-gray-300 rounded-lg bg-white py-2 shadow-lg"
-            >
-              {countries &&
-                Object.keys(countries).map((countryPath) => {
-                  const countryLocale = countries[countryPath];
-                  const isSelected =
-                    countryLocale.language === selectedLocale.language &&
-                    countryLocale.country === selectedLocale.country;
-
-                  return (
-                    <Popover.Close
-                      aria-label={`Select ${countryLocale.label} country`}
-                      key={countryPath}
-                      type="button"
-                      onClick={() =>
-                        handleLocaleChange({
-                          redirectTo: getCountryUrlPath({
-                            countryLocale,
-                            defaultLocalePrefix,
-                            pathWithoutLocale,
-                          }),
-                          buyerIdentity: {
-                            countryCode: countryLocale.country,
-                          },
-                        })
-                      }
-                      className="flex w-full cursor-pointer items-center gap-2 px-4 py-2 text-left text-body transition hover:bg-neutral-100"
-                    >
-                      <ReactCountryFlag
-                        svg
-                        countryCode={countryLocale.country}
-                        className="rounded-xs"
-                        style={{ width: "20px", height: "12px" }}
-                      />
-                      <span>{countryLocale.label}</span>
-                      {isSelected ? (
-                        <span className="ml-auto">
-                          <CheckCircleIcon className="size-5" />
-                        </span>
-                      ) : null}
-                    </Popover.Close>
-                  );
-                })}
-            </ScrollArea>
-          </Popover.Content>
-        </Popover.Portal>
-      </Popover.Root>
-      <div className="w-1 h-4 border-l border-gray-300" />
-    </div>
-  );
-}
